Move more rules out of typing category
