+++ fortitude/resources/test/fixtures/correctness/C061.f90
