+++ fortitude/resources/test/fixtures/correctness/C071.f90
