+++ fortitude/resources/test/fixtures/correctness/C072.f90
