+++ fortitude/src/rules/obsolescent/mod.rs
@@ -1,3 +1,4 @@
+pub mod assumed_size_character_syntax;
 pub mod common_blocks;
 pub mod computed_goto;
 pub mod entry_statement;
@@ -25,6 +26,7 @@ mod tests {
     #[test_case(Rule::SpecificName, Path::new("OB031.f90"))]
     #[test_case(Rule::ComputedGoTo, Path::new("OB041.f90"))]
     #[test_case(Rule::PauseStatement, Path::new("OB051.f90"))]
+    #[test_case(Rule::DeprecatedAssumedSizeCharacter, Path::new("OB061.f90"))]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.as_ref(), path.to_string_lossy());
         let diagnostics = test_path(
